backend/requirements.txt
fastapi
uvicorn
openai
numpy
pydantic
python-multipart
python-dotenv
